HUD displays player lives. Resolves issue #50.
Player ship entity now has a reference to the hud view to update player
lives count. Added a new type, playerbullet to differentiate between
colliding with the player or one of his or her bullets. Added a constant
variable START_LIVES that sets the starting lives for the player
(default is 3).