ESC now pauses gameplay
I have created a rudimentary method for pausing the game when the ESC
key is pressed. Right now all this really does is force the game to
prematurely exit the update function, but it is at least a start. I
mentioned this in the issue comments, but I'm hoping we can discuss what
we want done when the game pauses (I assume a in-game pause menu will be
involved). For issue #60.